@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { StateButtonComponent } from './state-button.component';
 
@@ -6,12 +6,12 @@ describe('StateButtonComponent', () => {
   let component: StateButtonComponent;
   let fixture: ComponentFixture<StateButtonComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(async () => {
     TestBed.configureTestingModule({
       declarations: [ StateButtonComponent ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(StateButtonComponent);
