@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
 import { NEVER } from 'rxjs';
 import { CollectionService } from 'src/app/core/services/collection.service';
@@ -10,11 +10,11 @@ describe('AddComponent', () => {
   let fixture: ComponentFixture<AddComponent>;
   let collectionServiceStub: Partial<CollectionService>;
 
-  beforeEach(async(() => {
+  beforeEach(async () => {
     collectionServiceStub = {
       collection: NEVER
     };
-    TestBed.configureTestingModule({
+    await TestBed.configureTestingModule({
       declarations: [ AddComponent ],
       imports: [
         FormsModule
@@ -27,7 +27,7 @@ describe('AddComponent', () => {
       ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(AddComponent);
